(ModbusSlave): ГЛОБАЛЬНОЕ ПЕРЕПИСЫВАНИЕ ФОРМАТОВ. Чтобы посылать в соответсвии со стандартом, посылать сообщение одним "пакетом" (по крайней мере одним write(buf,len), изменён формат ModbusMessage. А также проведён небольшой рефакторинг классов ModbusXXX.

+++ include/modbus/ModbusClient.h
@@ -170,12 +170,12 @@ class ModbusClient
 		    \param len     - size of buf
 		    \return real data lenght ( must be <= len )
 		*/
-		virtual size_t getNextData( unsigned char* buf, int len ) = 0;
+		virtual size_t getNextData( unsigned char* buf, size_t len ) = 0;
 
 		/*! set timeout for send/receive data */
 		virtual void setChannelTimeout( timeout_t msec ) = 0;
 
-		virtual ModbusRTU::mbErrCode sendData( unsigned char* buf, int len ) = 0;
+		virtual ModbusRTU::mbErrCode sendData( unsigned char* buf, size_t len ) = 0;
 
 		/*! функция запрос-ответ */
 		virtual ModbusRTU::mbErrCode query( ModbusRTU::ModbusAddr addr, ModbusRTU::ModbusMessage& msg,
