+++ include/modbus/ModbusTCPMaster.h
@@ -50,9 +50,9 @@ class ModbusTCPMaster:
 
 	protected:
 
-		virtual size_t getNextData( unsigned char* buf, int len ) override;
+		virtual size_t getNextData(unsigned char* buf, size_t len ) override;
 		virtual void setChannelTimeout( timeout_t msec ) override;
-		virtual ModbusRTU::mbErrCode sendData( unsigned char* buf, int len ) override;
+		virtual ModbusRTU::mbErrCode sendData( unsigned char* buf, size_t len ) override;
 		virtual ModbusRTU::mbErrCode query( ModbusRTU::ModbusAddr addr, ModbusRTU::ModbusMessage& msg,
 											ModbusRTU::ModbusMessage& reply, timeout_t timeout ) override;
